requirements.txt
langgraph>=0.2.0
langchain-google-genai>=1.0.0
google-cloud-bigquery>=3.13.0
pandas>=2.0.0
python-dotenv>=1.0.0
langchain_core>=0.3.0
db-dtypes==1.2.0
langchain-community>=0.2.0
tabulate>=0.9.0
pyyaml>=6.0
faiss-cpu>=1.7.4
